requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Django==4.2.16
environs==11.0.0
fonttools==4.55.0
idna==3.10
importlib_resources==6.4.5
kiwisolver==1.4.7
marshmallow==3.22.0
matplotlib==3.9.2
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scapy==2.6.0
scipy==1.13.1
six==1.16.0
sockets==1.0.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zipp==3.21.0
